requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
faster-whisper>=0.10.0
pyannote.audio>=3.3.2
torchaudio>=2.0.0
scipy>=1.10.0
